refactor(templates): split solara_map_app app.py into component packages
