+++ pysepal/templates/solara/solara_map_app/component/model/__init__.py
@@ -3,6 +3,8 @@
 This package contains the data models used by the map application template.
 """
 
+from .legend import LayerLegend
 from .model import AppModel
+from .processing import ProcessingOutputs
 
-__all__ = ["AppModel"]
+__all__ = ["AppModel", "LayerLegend", "ProcessingOutputs"]
