+++ pysepal/templates/solara/solara_map_app/component/widget/__init__.py
@@ -1,0 +1,6 @@
+"""Reusable widgets and hooks for the Solara map application."""
+
+from .legend import MapLegend
+from .map import use_aoi_scoped_layers, use_sepal_map
+
+__all__ = ["MapLegend", "use_aoi_scoped_layers", "use_sepal_map"]
